Move parseId functions to util.
