+++ rabbitmq/resource_federation_upstream_test.go
@@ -51,38 +51,6 @@ func TestAccFederationUpstream_hasComponent(t *testing.T) {
 	})
 }
 
-func TestParseId(t *testing.T) {
-	var badInputs = []string{
-		"",
-		"foo/test",
-		"footest",
-		"foo@bar@test",
-	}
-
-	for _, input := range badInputs {
-		_, _, err := parseId(input)
-		if err == nil {
-			t.Errorf("parseId failed for: %s.", input)
-		}
-	}
-
-	var goodInputs = []struct {
-		input string
-		name  string
-		vhost string
-	}{
-		{"foo@test", "foo", "test"},
-		{"foo@/", "foo", "/"},
-	}
-
-	for _, test := range goodInputs {
-		name, vhost, err := parseId(test.input)
-		if err != nil || name != test.name || vhost != test.vhost {
-			t.Errorf("parseId failed for: %s.", test.input)
-		}
-	}
-}
-
 func testAccFederationUpstreamCheck(rn string, upstream *rabbithole.FederationUpstream) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		rs, ok := s.RootModule().Resources[rn]
